// src/table/types/exmg-grid-smart-toolbar-types.ts
export type ActionWithCondition<TCondition extends BaseActionCondition = BaseActionCondition> = {
  id: string;
  icon?: string;
  text?: string;
  tooltip?: string;
  condition?: TCondition;
};

export enum ActionConditionType {
  AmountOfSelectedItemsRange = 'amount_of_items_range',
}

export interface BaseActionCondition {
  type?: ActionConditionType;
}

export interface ActionAmountSelectedItemsCondition extends BaseActionCondition {
  min?: number;
  max?: number;
}